# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ArcadeMatrix RPi Local API
  description: >
    Local HTTP API exposed by the ArcadeMatrix Raspberry Pi firmware (Actix-web,
    default port 8080). It drives the schema-driven Web UI: engines are
    auto-discovered from the registry (`GET /api/engines`), configured as
    generic instances (`/api/instances`), and sequenced by the rotation
    (`/api/rotation`). All config writes are self-healed (defaults injected,
    out-of-range values clamped/fallback) before being persisted atomically to
    `config.json`, and the running engine is notified live via
    `on_config_changed()`.
  version: 3.0.0
servers:
  - url: http://arcadematrix.local:8080
    description: Default local mDNS address
  - url: http://127.0.0.1:8080
    description: Local dev server (Mock matrix on Mac/Windows/Linux)
...
security:
  - ApiToken: []
paths:
...
  /api/update:
    post:
      tags: [Firmware]
      summary: OTA firmware upload
      description: >
        multipart/form-data upload of a compiled ELF binary. The payload is
        validated (ELF magic + matching CPU architecture) before replacing the
        running binary. Architecture mismatch is bypassed in debug builds only.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Update accepted; daemon restarts
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '400':
          description: Invalid firmware (bad magic, too small, or arch mismatch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/ota/auto-update:
    post:
      tags: [Firmware]
      summary: 1-Click Online OTA Download and Installation
      description: >
        Downloads target release binary from GitHub, validates ELF magic & architecture,
        atomically replaces binary at runtime and schedules service restart.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                download_url:
                  type: string
      responses:
        '200':
          description: Download complete, service restarting
        '400':
          description: Architecture mismatch or invalid download
        '502':
          description: Download network failure
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: >
        Required only when `api_auth_enabled` is true. Send the configured
        `api_token`. When auth is disabled, the header is ignored.
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
...